Export the `Serializable` type from the package.
This is needed to aide integration with other libraries. For instance,
if you want to make a generic function that will make use of an RpcStub,
you will need to declare a constraint of `Serializable<T>`, but you
currently can't because Serializable isn't exported.

The woraround is to copy the entire type definition and declare it
locally, which works, but this is much simpler.